% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brownfields_processed}
\alias{brownfields_processed}
\title{Brownfield Sites}
\format{
A data frame with 456 rows and 21 columns
\describe{
\item{cleanup_name}{The Name of the Property}
\item{location_address}{Street Address}
\item{city_name}{City}
\item{state_code}{State}
\item{postal_code}{ZIP Code}
\item{county_name}{County}
\item{epa_id}{EPA Identification Number}
\item{brownfields_link_csv}{Link to Brownfields Profile Page (if applicable)}
\item{rcra_csv}{Link to RCRA Profile Page (if applicable)}
\item{superfund_link_csv}{Link to Superfund Profile Page (if applicable)}
\item{echo_link_csv}{Link to ECHO Report Page (if applicable)}
\item{responselink_csv}{Link to Site Response Page (if applicable)}
\item{re_powering_csv}{Link to Re-Powering Site Profile Page (if applicable)}
\item{frs_link_csv}{Link to Facility Detail Report (if applicable)}
\item{map_site_csv}{Link to EPA Interactive Map (if applicable)}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Site Latitude}
\item{lon}{Site Longitude}
\item{frs}{Facility Registery Service Identification Number}
}
}
\source{
\url{https://ofmpub.epa.gov/apex/cimc/f?p=100:10::::::}
}
\usage{
brownfields_processed
}
\description{
A dataset containing an index of brownfield sites in South Carolina
NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
