% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contaminants_processed}
\alias{contaminants_processed}
\title{Contaminants of Concern}
\format{
A data frame with 1784 rows and 25 columns
\describe{
\item{contaminant_name}{Identified Contaminant}
\item{media}{Media Where Contaminant is Found}
\item{site_name}{The Name of the Site}
\item{epa_id}{EPA Identification Number}
\item{city}{City}
\item{state}{State}
\item{county}{County}
\item{zip_code}{ZIP Code}
\item{latitude}{Latitude}
\item{longitude}{Longitude}
\item{region}{EPA Region}
\item{npl_status}{National Priority List Status}
\item{superfund_alternative_agreement}{Superfund Alternative Agreement Indicator}
\item{federal_facility}{Federal Facility Indicator}
\item{operable_unit_number}{Portion of the NPL for Tracking and Planning}
\item{sequence_id}{Sequence Identifier}
\item{decision_document_type}{Record of Decision}
\item{actual_completion_date}{Completion Date of Sequence or Milestone}
\item{fiscal_year}{Fiscal Year}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.epa.gov/superfund/superfund-data-and-reports}
}
\usage{
contaminants_processed
}
\description{
The US EPA's records of contaminated sites and the contaminating substance
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
