% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_typology_codes}
\alias{county_typology_codes}
\title{SC County Typology Codes}
\format{
A data frame with 46 rows and 21 columns
\describe{
\item{fip_stxt}{5-Digit FIPS Code}
\item{state}{State}
\item{county_name}{County}
\item{metro_nonmetro_status_2013_0_nonmetro_1_metro}{Metro/Nonmetro (1 = Metro)}
\item{economic_types_type_2015_update_non_overlapping}{Number of Non-Overlapping Economic Types}
\item{economic_type_label}{Economic Type Label}
\item{farming_2015_update}{Farm-Dependent}
\item{mining_2015_update}{Mining-Dependent}
\item{manufacturing_2015_update}{Manufacturing-Dependent}
\item{government_2015_update}{Government-Dependent}
\item{recreation_2015_update}{Recreation-Dependent}
\item{nonspecialized_2015_update}{Nonspecialized}
\item{low_education_2015_update}{Low Education County}
\item{low_employment_cnty_2008_2012_25_64}{Low Employment County}
\item{pop_loss_2010}{Experienced Resident Decline}
\item{retirement_dest_2015_update}{Retirement Destination}
\item{persistent_poverty_2013}{Experienced Persistent Poverty}
\item{persistent_related_child_poverty_2013}{Experienced Persistent Related Child Poverty}
\item{geoid}{GEOID}
\item{county_latitude}{County Centroid Latitude}
\item{county_longitude}{County Centroid Longitude}
}
}
\source{
\url{https://www.ers.usda.gov/data-products/county-typology-codes/}
}
\usage{
county_typology_codes
}
\description{
The dataset of county social and economic characteristics, as well as a classification of economic dependence. A value of '1' indicates 'yes'
}
\keyword{datasets}
