% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{developed_land_imperviousness}
\alias{developed_land_imperviousness}
\title{Percent of Impervious Developed Land}
\format{
A data frame with 1103 rows and 9 columns
\describe{
\item{state_fips}{State 2-Digit FIPS Code}
\item{state}{State}
\item{geoid}{GEOID}
\item{census_tract_2}{County Name and Census Tract}
\item{year}{Data Collection Year}
\item{value}{Percentage of land that is impervious to water due to development in the census tract}
\item{data_comment}{NA}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
}
}
\source{
\url{https://ephtracking.cdc.gov/DataExplorer/}
}
\usage{
developed_land_imperviousness
}
\description{
The dataset of South Carolina census tracts with percentage estimates for the amount of land that is impervious to water due to development.
Impervious land increases precipitation runoff, as the water is not able to soak into the earth, and mobilizes pollutants and contaminants in
the runoff, allowing it an avenue to enter the waterways and spread its toxic effect. The processed dataset includes the census tract, value
estimate, and census information. The dataset is categorized as a Natural and Built Environment Determinant of Health.
}
\keyword{datasets}
