% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhomer-package.R
\docType{package}
\name{dhomer-package}
\alias{dhomer}
\alias{dhomer-package}
\title{dhomer: DHOME Ontology and Data}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Tami Crawford \email{tami.crawford@gmail.com} (\href{https://orcid.org/0000-0003-0720-7701}{ORCID})

Authors:
\itemize{
  \item Lauren Cuppy \email{cuppy.lauren@gmail.com}
}

}
\keyword{internal}
