% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fooddeserts_processed}
\alias{fooddeserts_processed}
\title{South Carolina Food Deserts}
\format{
A data frame with 1091 rows and 149 columns
\describe{
\item{census_tract}{Census tract}
\item{state}{State}
\item{county}{County}
\item{urban}{Urban tract}
\item{pop2010}{Population, tract total}
\item{ohu2010}{Housing units, total}
\item{group_quarters_flag}{Group quarters, tract with high share}
\item{numgqtrs}{Group quarters, tract population residing in, number}
\item{pctgqtrs}{Group quarters, tract population residing in, share}
\item{lila_tracts_1and10}{Low income and low access tract measured at 1 mile for urban areas and 10 miles for rural areas}
\item{lila_tracts_halfand10}{Low income and low access tract measured at 1/2 mile for urban areas and 10 miles for rural areas}
\item{lila_tracts_1and20}{Low income and low access tract measured at 1 mile for urban areas and 20 miles for rural areas}
\item{lila_tracts_vehicle}{Low income and low access tract using vehicle access or low income and low access tract measured at 20 miles}
\item{hunv_flag}{Vehicle access, tract with low vehicle access}
\item{low_income_tracts}{Low income tract}
\item{poverty_rate}{Tract poverty rate}
\item{median_family_income}{Tract median family income}
\item{la1and10}{Low access tract at 1 mile for urban areas and 10 miles for rural areas}
\item{l_ahalfand10}{Low access tract at 1/2 mile for urban areas and 10 miles for rural areas}
\item{la1and20}{Low access tract at 1 mile for urban areas and 20 miles for rural areas}
\item{la_tracts_half}{Low access tract at 1/2 mile}
\item{la_tracts1}{Low access tract at 1 mile}
\item{la_tracts10}{Low access tract at 10 miles}
\item{la_tracts20}{Low access tract at 20 miles}
\item{la_tracts_vehicle_20}{Low access tract using vehicle access and at 20 miles  in rural areas}
\item{lapop1_10}{Low access, population at 1 mile for urban areas and 10 miles for rural areas, number}
\item{lapop05_10}{Low access, population at 1/2 mile for urban areas and 10 miles for rural areas, number}
\item{lapop1_20}{Low access, population at 1 mile for urban areas and 20 miles for rural areas, number}
\item{lalowi1_10}{Low access, low-income population at 1 mile for urban areas and 10 miles for rural areas, number}
\item{lalowi05_10}{Low access, low-income population at 1/2 mile for urban areas and 10 miles for rural areas, number}
\item{lalowi1_20}{Low access, low-income population at 1 mile for urban areas and 20 miles for rural areas, number}
\item{lapophalf}{Low access, population at 1/2 mile, number}
\item{lapophalfshare}{Low access, population at 1/2 mile, share}
\item{lalowihalf}{Low access, low-income population at 1/2 mile, number}
\item{lalowihalfshare}{Low access, low-income population at 1/2 mile, share}
\item{lakidshalf}{Low access, children age 0-17 at 1/2 mile, number}
\item{lakidshalfshare}{Low access, children age 0-17 at 1/2 mile, share}
\item{laseniorshalf}{Low access, seniors age 65+ at 1/2 mile, number}
\item{laseniorshalfshare}{Low access, seniors age 65+ at 1/2 mile, share}
\item{lawhitehalf}{Low access, White population at 1/2 mile, number}
\item{lawhitehalfshare}{Low access, White population at 1/2 mile, share}
\item{lablackhalf}{Low access, Black or African American population at 1/2 mile, number}
\item{lablackhalfshare}{Low access, Black or African American population at 1/2 mile, share}
\item{laasianhalf}{Low access, Asian population at 1/2 mile, number}
\item{laasianhalfshare}{Low access, Asian population at 1/2 mile, share}
\item{lanhopihalf}{Low access, Native Hawaiian or Other Pacific Islander population at 1/2 mile, number}
\item{lanhopihalfshare}{Low access, Native Hawaiian or Other Pacific Islander population at 1/2 mile, share}
\item{laaianhalf}{Low access, American Indian or Alaska Native population at 1/2 mile, number}
\item{laaianhalfshare}{Low access, American Indian or Alaska Native population at 1/2 mile, share}
\item{laomultirhalf}{Low access, Other/Multiple race population at 1/2 mile, number}
\item{laomultirhalfshare}{Low access, Other/Multiple race population at 1/2 mile, share}
\item{lahisphalf}{Low access, Hispanic or Latino population at 1/2 mile, number}
\item{lahisphalfshare}{Low access, Hispanic or Latino population at 1/2 mile, share}
\item{lahunvhalf}{Vehicle access, housing units without and low access at 1/2 mile, number}
\item{lahunvhalfshare}{Vehicle access, housing units without and low access at 1/2 mile, share}
\item{lasnaphalf}{Low access, housing units receiving SNAP benefits at 1/2 mile, number}
\item{lasnaphalfshare}{Low access, housing units receiving SNAP benefits at 1/2 mile, share}
\item{lapop1}{Low access, population at 1 mile, number}
\item{lapop1share}{Low access, population at 1 mile, share}
\item{lalowi1}{Low access, low-income population at 1 mile, number}
\item{lalowi1share}{Low access, low-income population at 1 mile, share}
\item{lakids1}{Low access, children age 0-17 at 1 mile, number}
\item{lakids1share}{Low access, children age 0-17 at 1 mile, share}
\item{laseniors1}{Low access, seniors age 65+ at 1 mile, number}
\item{laseniors1share}{Low access, seniors age 65+ at 1 mile, share}
\item{lawhite1}{Low access, White population at 1 mile, number}
\item{lawhite1share}{Low access, White population at 1 mile, share}
\item{lablack1}{Low access, Black or African American population at 1 mile, number}
\item{lablack1share}{Low access, Black or African American population at 1 mile, share}
\item{laasian1}{Low access, Asian population at 1 mile, number}
\item{laasian1share}{Low access, Asian population at 1 mile, share}
\item{lanhopi1}{Low access, Native Hawaiian and Other Pacific Islander population at 1 mile, number}
\item{lanhopi1share}{Low access, Native Hawaiian and Other Pacific Islander population at 1 mile, share}
\item{laaian1}{Low access, American Indian and Alaska Native population at 1 mile, number}
\item{laaian1share}{Low access, American Indian and Alaska Native population at 1 mile, share}
\item{laomultir1}{Low access, Other/Multiple race population at 1 mile, number}
\item{laomultir1share}{Low access, Other/Multiple race population at 1 mile, share}
\item{lahisp1}{Low access, Hispanic or Latino population at 1 mile, number}
\item{lahisp1share}{Low access, Hispanic or Latino population at 1 mile, share}
\item{lahunv1}{Vehicle access, housing units without and low access at 1 mile, number}
\item{lahunv1share}{Vehicle access, housing units without and low access at 1 mile, share}
\item{lasnap1}{Low access, housing units receiving SNAP benefits at 1 mile, number}
\item{lasnap1share}{Low access, housing units receiving SNAP benefits at 1 mile, share}
\item{lapop10}{Low access, population at 10 miles, number}
\item{lapop10share}{Low access, population at 10 miles, share}
\item{lalowi10}{Low access, low-income population at 10 miles, number}
\item{lalowi10share}{Low access, low-income population at 10 miles, share}
\item{lakids10}{Low access, children age 0-17 at 10 miles, number}
\item{lakids10share}{Low access, children age 0-17 at 10 miles, share}
\item{laseniors10}{Low access, seniors age 65+ at 10 miles, number}
\item{laseniors10share}{Low access, seniors age 65+ at 10 miles, share}
\item{lawhite10}{Low access, White population at 10 miles, number}
\item{lawhite10share}{Low access, White population at 10 miles, share}
\item{lablack10}{Low access, Black or African American population at 10 miles, number}
\item{lablack10share}{Low access, Black or African American population at 10 miles, share}
\item{laasian10}{Low access, Asian population at 10 miles, number}
\item{laasian10share}{Low access, Asian population at 10 miles, share}
\item{lanhopi10}{Low access, Native Hawaiian and Other Pacific Islander population at 10 miles, number}
\item{lanhopi10share}{Low access, Native Hawaiian and Other Pacific Islander population at 10 miles, share}
\item{laaian10}{Low access, American Indian and Alaska Native population at 10 miles, number}
\item{laaian10share}{Low access, American Indian and Alaska Native population at 10 miles, share}
\item{laomultir10}{Low access, Other/Multiple race population at 10 miles, number}
\item{laomultir10share}{Low access, Other/Multiple race population at 10 miles, share}
\item{lahisp10}{Low access, Hispanic or Latino population at 10 miles, number}
\item{lahisp10share}{Low access, Hispanic or Latino population at 10 miles, share}
\item{lahunv10}{Vehicle access, housing units without and low access at 10 miles, number}
\item{lahunv10share}{Vehicle access, housing units without and low access at 10 miles, share}
\item{lasnap10}{Low access, housing units receiving SNAP benefits at 10 miles, number}
\item{lasnap10share}{Low access,housing units receiving SNAP benefits at 10 miles, share}
\item{lapop20}{Low access, population at 20 miles, number}
\item{lapop20share}{Low access, population at 20 miles, share}
\item{lalowi20}{Low access, low-income population at 20 miles, number}
\item{lalowi20share}{Low access, low-income population at 20 miles, share}
\item{lakids20}{Low access, children age 0-17 at 20 miles, number}
\item{lakids20share}{Low access, children age 0-17 at 20 miles, share}
\item{laseniors20}{Low access, seniors age 65+ at 20 miles, number}
\item{laseniors20share}{Low access, seniors age 65+ at 20 miles, share}
\item{lawhite20}{Low access, White population at 20 miles, number}
\item{lawhite20share}{Low access, White population at 20 miles, share}
\item{lablack20}{Low access, Black or African American population at 20 miles, number}
\item{lablack20share}{Low access, Black or African American population at 20 miles, share}
\item{laasian20}{Low access, Asian population at 20 miles, number}
\item{laasian20share}{Low access, Asian population at 20 miles, share}
\item{lanhopi20}{Low access, Native Hawaiian and Other Pacific Islander population at 20 miles, number}
\item{lanhopi20share}{Low access, Native Hawaiian and Other Pacific Islander population at 20 miles, share}
\item{laaian20}{Low access, American Indian and Alaska Native population at 20 miles, number}
\item{laaian20share}{Low access, American Indian and Alaska Native population at 20 miles, share}
\item{laomultir20}{Low access, Other/Multiple race population at 20 miles, number}
\item{laomultir20share}{Low access, Other/Multiple race population at 20 miles, share}
\item{lahisp20}{Low access, Hispanic or Latino population at 20 miles, number}
\item{lahisp20share}{Low access, Hispanic or Latino population at 20 miles, share}
\item{lahunv20}{Vehicle access, housing units without and low access at 20 miles, number}
\item{lahunv20share}{Vehicle access, housing units without and low access at 20 miles, share}
\item{lasnap20}{Low access, housing units receiving SNAP benefits at 20 miles, number}
\item{lasnap20share}{Low access, housing units receiving SNAP benefits at 20 miles, share}
\item{tractlowi}{Tract low-income population, number}
\item{tractkids}{Tract children age 0-17, number}
\item{tractseniors}{Tract seniors age 65+, number}
\item{tractwhite}{Tract White population, number}
\item{tractblack}{Tract Black or African American population, number}
\item{tractasian}{Tract Asian population, number}
\item{tractnhopi}{Tract Native Hawaiian and Other Pacific Islander population, number}
\item{tractaian}{Tract American Indian and Alaska Native population, number}
\item{tractomultir}{Tract Other/Multiple race population, number}
\item{tracthispanic}{Tract Hispanic or Latino population, number}
\item{tracthunv}{Tract housing units without a vehicle, number}
\item{tractsnap}{Tract housing units receiving SNAP benefits, number}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}

}
}
\source{
\url{https://www.ers.usda.gov/data-products/food-access-research-atlas/download-the-data/}
}
\usage{
fooddeserts_processed
}
\description{
The dataset of census tracts deemed to be food deserts by the USDA
}
\keyword{datasets}
