% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{free_clinics}
\alias{free_clinics}
\title{South Carolina Free Clinics}
\format{
A data frame with 70 rows and 8 columns
\describe{
\item{clinic_name}{The Name of the Clinic}
\item{specialty}{Clinic Specialty}
\item{hours_of_operation}{Hours of Operation}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract CentroidLongitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.scfreeclinics.org/clinics/}
}
\usage{
free_clinics
}
\description{
The dataset of the free medical clinics in South Carolina
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
