% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospitals}
\alias{hospitals}
\title{South Carolina Hospitals}
\format{
A data frame with 112 rows and 9 columns
\describe{
\item{id}{ID Number}
\item{name}{Hospital Name}
\item{telephone}{Phone Number}
\item{trauma}{Trauma Center Level}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://hifld-geoplatform.opendata.arcgis.com/datasets/hospitals/explore}
}
\usage{
hospitals
}
\description{
The dataset of hospitals in South Carolina
}
\keyword{datasets}
