% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hrsa}
\alias{hrsa}
\title{All HRSA Sites}
\format{
A data frame with 252 rows and 9 variables:
\describe{
\item{site_name}{The site name}
\item{bhcmis_org_id}{BHCMIS Organization ID}
\item{bphc_assigned_number}{BPHC Assigned Number}
\item{site_telephone_number}{Site Telephone Number}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Latitude}
\item{tract_longitude}{Tract Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://data.hrsa.gov/data/download/}
}
\usage{
hrsa
}
\description{
A dataset containing HRSA sites
NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
