% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internet_access}
\alias{internet_access}
\title{Percent of SC Households without Internet Access}
\format{
A data frame with 1103 rows and 10 columns
\describe{
\item{state_fips}{State FIPS Code}
\item{state}{State}
\item{geoid}{GEOID}
\item{census_tract_2}{County and Census Tract}
\item{year}{Data Collection Year}
\item{value}{Percentage Value}
\item{data_comment}{NA}
\item{x8}{NA}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
}
}
\source{
\url{https://ephtracking.cdc.gov/DataExplorer/}
}
\usage{
internet_access
}
\description{
The dataset of census tracts in South Carolina with percentage estimates of households without internet access
}
\keyword{datasets}
