% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{libraries}
\alias{libraries}
\title{South Carolina Public Libraries}
\format{
A data frame with 232 rows and 9 columns
\describe{
\item{libid}{Library ID}
\item{libname}{Name}
\item{phone}{Phone Number}
\item{hours}{Public Service Hours Per Year}
\item{geoid}{GEOID}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.imls.gov/research-evaluation/data-collection/public-libraries-survey}
}
\usage{
libraries
}
\description{
The dataset of public libraries and library outlets in South Carolina. Public libraries are an accessible collection of books and other
resources for the community. The processed dataset includes the library name, identification numbers, telephone numbers, hours open,
latitude/longitude, and census information. The dataset is categorized as an Education Determinant of Health.
}
\keyword{datasets}
