% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markets}
\alias{markets}
\title{South Carolina Farmers and Roadside Markets}
\format{
A data frame with 337 rows and 8 columns
\describe{
\item{objectid}{Object ID, which SC DHEC uses for identification of sites when mapping with ArcGIS}
\item{name}{The Name of the Market}
\item{mail_phone}{Phone Number}
\item{geoid}{GEOID}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://sc-department-of-health-and-environmental-control-gis-sc-dhec.hub.arcgis.com/datasets/cdb563626aaf4477b7ea591fec10a3e2_0/explore}
}
\usage{
markets
}
\description{
The dataset of the farmer's and roadside markets in South Carolina, compiled by SC DHEC. These markets provide seasonal access to local
produce to their communities, which tends to be harder to access especially in communities that lack access to a chain supermarket store.
The processed dataset includes the market name, phone number, latitude/longitude, and census tract information. The dataset is categorized
as a Social and Community Context Determinant of Health.
}
\details{
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
