% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markets_processed}
\alias{markets_processed}
\title{South Carolina Farmers and Roadside Markets}
\format{
A data frame with 337 rows and 38 columns
\describe{
\item{x}{X Coordinate}
\item{y}{Y Coordinate}
\item{objectid}{Object ID}
\item{name}{The Name of the Market}
\item{street}{Street Address}
\item{city}{City}
\item{state}{State}
\item{zip}{ZIP Code}
\item{county}{County}
\item{mail_phone}{Phone Number}
\item{website}{Website}
\item{open_jan}{Open in January}
\item{open_feb}{Open in February}
\item{open_mar}{Open in March}
\item{open_apr}{Open in April}
\item{open_may}{Open in May}
\item{open_jun}{Open in June}
\item{open_jul}{Open in July}
\item{open_aug}{Open in August}
\item{open_sep}{Open in September}
\item{open_oct}{Open in October}
\item{open_nov}{Open in November}
\item{open_dec}{Open in December}
\item{cash}{Accepts Cash Payment}
\item{check}{Accepts Check Payment}
\item{credit}{Accepts Credit Payment}
\item{debit}{Accepts Debit Payment}
\item{other_pay}{Accepts Other Forms of Payment}
\item{wic}{Accepts WIC Payments}
\item{senior}{Accepts Senior Vouchers}
\item{snap_ebt}{Accepts SNAP/EBT Payment}
\item{snap_trans}{SNAP Transaction Type}
\item{h_bucks}{Accepts Healthy Bucks Payment}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://sc-department-of-health-and-environmental-control-gis-sc-dhec.hub.arcgis.com/datasets/cdb563626aaf4477b7ea591fec10a3e2_0/explore}
}
\usage{
markets_processed
}
\description{
The dataset of the farmers and roadside markets in South Carolina
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
