% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pharmacies}
\alias{pharmacies}
\title{South Carolina Pharmacy Network}
\format{
A data frame with 954 rows and 10 columns
\describe{
\item{npi}{National Provider Identifier Number}
\item{basic_organization_name}{The Name of the Pharmacy Organization}
\item{hours_of_operation}{Hours of Operation}
\item{medicaid_id}{Medicaid Identifier}
\item{addresses_telephone_number}{Telephone Number}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.selecthealthofsc.com/pdf/provider/pharmacy/pharmacy-network.pdf}
}
\usage{
pharmacies
}
\description{
The dataset of the pharmacies in South Carolina
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
