% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pharmacies_processed}
\alias{pharmacies_processed}
\title{South Carolina Pharmacy Network}
\format{
A data frame with 954 rows and 54 columns
\describe{
\item{npi}{National Provider Identifier Number}
\item{basic_organization_name}{The Name of the Pharmacy Organization}
\item{npi_2}{National Provider Identifier Number}
\item{hours_of_operation}{Hours of Operation}
\item{medicaid_id}{Medicaid Identifier}
\item{spoken_languages}{Spoken Languages}
\item{e_rx}{Offers e-Rx}
\item{delivery}{Offers Delivery}
\item{compound}{Offers Compounds}
\item{drive_thru}{Offers Drive-Thru}
\item{vaccines}{Offers Vaccines}
\item{handicap_access}{Offers Handicap Access}
\item{cultural_comp_training}{Provides Cultural Competence Training}
\item{accepting_new_customers}{Accepting New Customers}
\item{website}{Website}
\item{basic_organizational_subpart}{Organizational Subpart?}
\item{basic_enumeration_date}{Enumeration Date}
\item{basic_last_updated}{Date Last Updated}
\item{basic_status}{Status (Active)}
\item{basic_authorized_official_credential}{Authorized Official Credential}
\item{basic_authorized_official_first_name}{Authorized Official First Name}
\item{basic_authorized_official_last_name}{Authorized Official Last Name}
\item{basic_authorized_official_telephone_number}{Authorized OFficial Telephone Number}
\item{basic_authorized_official_title_or_position}{Authorized OFficial Title/Position}
\item{basic_name}{Pharmacy Name}
\item{basic_parent_organization_ein}{Parent Organization EIN}
\item{basic_parent_organization_legal_business_name}{Parent Organization}
\item{basic_authorized_official_middle_name}{Authorized Official Middle Name}
\item{basic_authorized_official_name_prefix}{Authorized Official Name Prefix}
\item{basic_authorized_official_name_suffix}{Authorized Official Name Suffix}
\item{basic_certification_date}{Certification Date}
\item{basic_name_prefix}{Other Official Name Prefix}
\item{basic_first_name}{Other Official First Name}
\item{basic_last_name}{Other Official Last Name}
\item{basic_middle_name}{Other Official Middle Name}
\item{basic_credential}{Other Official redentials}
\item{basic_sole_proprieter}{Sole Proprietor}
\item{basic_gender}{Other Official Gender}
\item{addresses_country_code}{Country Code}
\item{addresses_country_name}{Country}
\item{addresses_address_purpose}{Address Purpose}
\item{addresses_address_type}{Address Type}
\item{addresses_address_1}{Street Address}
\item{addresses_address_2}{Street Address 2}
\item{addresses_city}{City}
\item{addresses_state}{State}
\item{addresses_postal_code}{ZIP Code}
\item{addresses_telephone_number}{Telephone Number}
\item{addresses_fax_number}{Fax Number}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.selecthealthofsc.com/pdf/provider/pharmacy/pharmacy-network.pdf}
}
\usage{
pharmacies_processed
}
\description{
The dataset of the pharmacies in South Carolina
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
