% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{property_crime}
\alias{property_crime}
\title{South Carolina Property Crime Statistics}
\format{
A data frame with 46 rows and 7 columns
\describe{
\item{county}{County}
\item{number}{Number of Crimes}
\item{rate}{Crime Rate}
\item{clearance}{Clearance Rate}
\item{geoid}{GEOID}
\item{county_latitude}{County Latitude}
\item{county_longitude}{County Longitude}
}
}
\source{
\url{https://www.sled.sc.gov/forms/statistics/2019\%20Crime\%20in\%20South\%20Carolina.pdf}
}
\usage{
property_crime
}
\description{
The dataset of property crime statistics in South Carolina in 2019 according to the South Carolina Law Enforcement Division
}
\keyword{datasets}
