% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protected_areas}
\alias{protected_areas}
\title{SC Protected Areas}
\format{
A data frame with 1191565 rows and 7 columns
\describe{
\item{category}{Category}
\item{unit_nm}{Unit Name}
\item{GEOID}{GEOID}
\item{tract_latitude}{Tract Latitude}
\item{tract_longitude}{Tract Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.usgs.gov/core-science-systems/science-analytics-and-synthesis/gap/science/pad-us-data-overview?qt-science_center_objects=4#qt-science_center_objects}
}
\usage{
protected_areas
}
\description{
The dataset of South Carolina's protected lands and waters
}
\keyword{datasets}
