% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public_schools}
\alias{public_schools}
\title{South Carolina Public Schools}
\format{
A data frame with 1183 rows and 9 columns
\describe{
\item{ncessch}{NCES school ID}
\item{leaid}{Unique Agency Code}
\item{leanm05}{Operating Agency}
\item{schnam05}{School Name}
\item{geoid}{GEOID}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://nces.ed.gov/ccd/CCDLocaleCode.asp}
}
\usage{
public_schools
}
\description{
The dataset of South Carolina public elementary and secondary schools. Proximity to the school may be a determinant in school completion and,
thus, higher educational attainment, especially for households without access to a vehicle. Similarly, because a great deal of time is spent
in school, proximity of the school to other determinants of health should also be considered. The processed dataset includes the school name,
identification numbers, latitude/longtiude, and census information. The dataset is categorized as an Education Determinant of Health.
}
\keyword{datasets}
