% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{social_vulnerability_index}
\alias{social_vulnerability_index}
\title{CDC/ATSDR Social Vulnerability Index}
\format{
A data frame with 1103 rows and 4 columns
\describe{
\item{geoid}{GEOID}
\item{svi_value}{Social Vulnerability Value}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
}
}
\source{
\url{https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html}
}
\usage{
social_vulnerability_index
}
\description{
The dataset of
}
\keyword{datasets}
