% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{superfund}
\alias{superfund}
\title{Superfund Sites}
\format{
A data frame with 33 rows and 9 columns
\describe{
\item{cleanup_name}{The Name of the Property}
\item{epa_id}{EPA Identification Number}
\item{superfund_id}{Superfund ID Number}
\item{frs}{Facility Registery Service Identification Number}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Site Latitude}
\item{lon}{Site Longitude}
\item{frs}{Facility Registery Service Identification Number}
}
}
\source{
\url{https://ofmpub.epa.gov/apex/cimc/f?p=100:10::::::}
}
\usage{
superfund
}
\description{
A dataset containing an index of US EPA Superfund sites in South Carolina
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
