% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transit_terminals}
\alias{transit_terminals}
\title{South Carolina Intermodal Passenger Transportation Terminals}
\format{
A data frame with 104 rows and 13 columns
\describe{
\item{fac_id}{Facility ID}
\item{fac_name}{Facility Name}
\item{notes}{Notes}
\item{mode_bus}{Provides Bus Services}
\item{mode_air}{Provides Air Services}
\item{mode_rail}{Provides Rail Services}
\item{mode_ferry}{Provides Ferry Services}
\item{mode_bike}{Provides Bike Services}
\item{geoid}{GEOID}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://hifld-geoplatform.opendata.arcgis.com/datasets/intermodal-passenger-connectivity-database-ipcd/explore?location=45.502900\%2C58.184500\%2C3.70}
}
\usage{
transit_terminals
}
\description{
The dataset of South Carolina's passenger transportation terminals that provide various modes of transportation. Public transit is essential
for those without access to a vehicle and those living areas with a low walkability index. The dataset classifies the transit terminals as
providing bus, air, rail, ferry, or bike transportation, as well as includes the facility name, identification number, latitude/longitude,
and census information. The dataset is categorized as a Natural and Built Environment Determinant of Health.
}
\keyword{datasets}
