% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tri}
\alias{tri}
\title{SC Toxic Release Inventory Form R and A Submissions}
\format{
A data frame with 59834 rows and 17 columns
\describe{
\item{chem_name}{Chemical Released}
\item{facility_name}{Facility Name}
\item{tri_chem_id}{TRI Chemical ID}
\item{tri_facility_id}{TRI Facility ID}
\item{naics_codes}{NAICS Code}
\item{total_off_site_release}{Total Pounds of Off-Site Toxic Chemical Release}
\item{total_on_off_site_release}{Total Pounds of On- and Off-Site Toxic Chemical Release}
\item{total_on_site_release}{Total Pounds of On-Site Toxic Chemical Release}
\item{total_production_related_waste}{Total Pounds of Toxic Chemicals Released due to Production}
\item{land_total_release}{Total Pounds of Toxic Chemicals Released to Land}
\item{air_total_release}{Total Pounds of Toxic Chemicals Released to Air}
\item{water_total_release}{Total Pounds of Toxic Chemicals Released to Water}
\item{geoid}{GEOID}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://enviro.epa.gov/facts/tri/form_ra_download.html}
}
\usage{
tri
}
\description{
The dataset of sites in South Carolina that have submitted a Toxic Release Inventory (TRI) Form R or A. The TRI tracks the management of
hazardous chemicals which pose a threat to the environment or human health. The program collects annual data from sites that manage hazardous
chemicals and requires the reporting of the amount of toxic chemicals released into the environment via land, water, and air. Form R is
required for sites that: utilize a PBT chemical (such as lead), manufacture, process, or use more than 1,000,000 pounds of a toxic chemical,
or has an annual reportable amount of release, recycling, energy recovery, or treatment of a toxic chemical that is more than 500 pounds.
Form A is to be filled out for: non-PBT chemicals, use of less than 1,000,000 pounds of a toxic chemical, and a annual reportable amount of
less than 500 pounds of a toxic chemical.  The dataset includes the chemical name, facility name, identifications, toxic chemical release
data (in pounds), latitude/longitude, and census tract information. The dataset is categorized as a Natural and Built Environment
Determinant of Health.
}
\keyword{datasets}
