% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicle_availability}
\alias{vehicle_availability}
\title{Percent of Housing Units with No Vehicle Access}
\format{
A data frame with 1103 rows and 9 columns
\describe{
\item{state_fips}{State 2-digit FIPS Code}
\item{state}{State}
\item{geoid}{GEOID}
\item{census_tract_2}{County Name and Census Tract}
\item{year}{Data Collection Year}
\item{value}{Percentage of housing units that do not have access to a vehicle in the census tract}
\item{data_comment}{NA}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
}
}
\source{
\url{https://ephtracking.cdc.gov/DataExplorer/}
}
\usage{
vehicle_availability
}
\description{
The dataset of South Carolina census tracts with percentage estimates of housing units with no vehicle access. Vehicle access plays a direct
role in an individual's ability to access food, community resources, and healthcare, especially in areas that are not highly walkable. The
processed dataset includes the census tract, the value estimate, and census information. The dataset is categorized as a Social and Community
Context Determinant of Health.
}
\keyword{datasets}
