% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brownfields}
\alias{brownfields}
\title{Brownfield Sites}
\format{
A data frame with 456 rows and 8 columns
\describe{
\item{cleanup_name}{The Name of the Property}
\item{brownfields_id}{Brownfields ID}
\item{frs_id}{Facility Registry ID}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Site Latitude}
\item{lon}{Site Longitude}
}
}
\source{
\url{https://ofmpub.epa.gov/apex/cimc/f?p=100:10::::::}
}
\usage{
brownfields
}
\description{
A dataset containing an index of brownfield sites in South Carolina
NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
