% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contaminants}
\alias{contaminants}
\title{Contaminants of Concern}
\format{
A data frame with 1784 rows and 9 columns
\describe{
\item{contaminant_name}{Identified Contaminant}
\item{media}{Media Where Contaminant is Found}
\item{site_name}{The Name of the Site}
\item{epa_id}{EPA Identification Number}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.epa.gov/superfund/superfund-data-and-reports}
}
\usage{
contaminants
}
\description{
The US EPA's records of contaminated sites and the contaminating substance
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
