% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{developed_land_imperviousness}
\alias{developed_land_imperviousness}
\title{Percent of Impervious Developed Land}
\format{
A data frame with 1103 rows and 9 columns
\describe{
\item{state_fips}{State FIPS Code}
\item{state}{State}
\item{geoid}{GEOID}
\item{census_tract_2}{County and Census Tract}
\item{year}{Data Collection Year}
\item{value}{Percentage Value}
\item{data_comment}{NA}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
}
}
\source{
\url{https://ephtracking.cdc.gov/DataExplorer/}
}
\usage{
developed_land_imperviousness
}
\description{
The dataset of South Carolina census tracts with percentage estimates for the amount impervious land in the tract due to development
}
\keyword{datasets}
