% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhomer-package.R
\docType{package}
\name{dhomer-package}
\alias{dhomer}
\alias{dhomer-package}
\title{dhomer: DHOME Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tamilyn.github.io/dhomer}
}

}
\author{
\strong{Maintainer}: Tami Crawford \email{tami.crawford@gmail.com} (\href{https://orcid.org/0000-0003-0720-7701}{ORCID})

Authors:
\itemize{
  \item Alexander Alekseyenko
  \item Lauren Cuppy \email{cuppy.lauren@gmail.com} (\href{https://orcid.org/0000-0002-1672-4260}{ORCID})
}

}
\keyword{internal}
