% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_deserts}
\alias{food_deserts}
\title{South Carolina Food Deserts}
\format{
A data frame with 1091 rows and 10 columns
\describe{
\item{census_tract}{Census tract}
\item{lila_tracts_1and10}{Low income and low access tract measured at 1 mile for urban areas and 10 miles for rural areas}
\item{lila_tracts_half_and10}{Low income and low access tract measured at 1/2 mile for urban areas and 10 miles for rural areas}
\item{lila_tracts_1and20}{Low income and low access tract measured at 1 mile for urban areas and 20 miles for rural areas}
\item{low_income_tracts}{Low income tract}
\item{poverty_rate}{Tract poverty rate}
\item{median_family_income}{Tract median family income}
\item{GEOID}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}

}
}
\source{
\url{https://www.ers.usda.gov/data-products/food-access-research-atlas/download-the-data/}
}
\usage{
food_deserts
}
\description{
The dataset of census tracts deemed to be food deserts by the USDA
}
\keyword{datasets}
