% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health_facilities}
\alias{health_facilities}
\title{South Carolina Health Facilities}
\format{
A data frame with 2781 rows and 7 columns
\describe{
\item{permit_type}{Type of Permit}
\item{name_of_facility}{Facility Name}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://sc-department-of-health-and-environmental-control-gis-sc-dhec.hub.arcgis.com/datasets/04d86b6899fc450288ba07d0dcab2d73_0/explore}
}
\usage{
health_facilities
}
\description{
A dataset of licensed health facilities as compiled by SC DHEC
}
\keyword{datasets}
