% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrsa.R
\docType{data}
\name{hrsa}
\alias{hrsa}
\title{All HRSA Sites}
\format{
A data frame with 252 rows and 31 variables:
\describe{
\item{site_name}{The site name}
\item{services_delivered_at_site}{to be filled out soon}
\item{health_center_type}{to be filled out soon}
\item{health_center_location_type}{to be filled out soon}
\item{health_center_location_setting}{to be filled out soon}
\item{county}{to be filled out soon}
\item{state}{to be filled out soon}
\item{grant_number}{to be filled out soon}
\item{bhcmis_org_id}{to be filled out soon}
\item{bphc_assigned_number}{to be filled out soon}
\item{health_center_look_alike_number}{to be filled out soon}
\item{address}{to be filled out soon}
\item{city}{to be filled out soon}
\item{zip}{to be filled out soon}
\item{site_telephone_number}{to be filled out soon}
\item{website}{to be filled out soon}
\item{admin_functions_performed_at_site}{to be filled out soon}
\item{hrsa_grant_programs}{to be filled out soon}
\item{native_hawaiian_health_center_admin_site}{to be filled out soon}
\item{health_center_service_delivery_site_group}{to be filled out soon}
\item{hhs_region}{to be filled out soon}
\item{u_s_senator_name_1}{to be filled out soon}
\item{u_s_senator_name_2}{to be filled out soon}
\item{cong_district}{to be filled out soon}
\item{u_s_congressional_representative_name}{to be filled out soon}
\item{in_u_s_mex_border_100_km_area}{to be filled out soon}
\item{in_u_s_mex_border_county}{to be filled out soon}
\item{rural_status}{to be filled out soon}
\item{longitude}{to be filled out soon}
\item{latitude}{to be filled out soon}
\item{grantee_name}{to be filled out soon}
}
}
\source{
\url{https://data.hrsa.gov/data/download/}
}
\usage{
hrsa
}
\description{
A dataset containing HRSA sites
NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
