% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{libraries}
\alias{libraries}
\title{South Carolina Public Libraries}
\format{
A data frame with 232 rows and 9 columns
\describe{
\item{libid}{Library ID}
\item{libname}{Name}
\item{phone}{Phone Number}
\item{hours}{Public Service Hours Per Year}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.imls.gov/research-evaluation/data-collection/public-libraries-survey}
}
\usage{
libraries
}
\description{
The dataset of public libraries and outlets in South Carolina
}
\keyword{datasets}
