% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markets}
\alias{markets}
\title{South Carolina Farmers and Roadside Markets}
\format{
A data frame with 337 rows and 8 columns
\describe{
\item{objectid}{Object ID}
\item{name}{The Name of the Market}
\item{mail_phone}{Phone Number}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://sc-department-of-health-and-environmental-control-gis-sc-dhec.hub.arcgis.com/datasets/cdb563626aaf4477b7ea591fec10a3e2_0/explore}
}
\usage{
markets
}
\description{
The dataset of the farmers and roadside markets in South Carolina
#' NEED TO CHECK for any requirements in redistributing data such as a notice
copyright, etc
}
\keyword{datasets}
