% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protected_areas}
\alias{protected_areas}
\title{SC Protected Areas}
\format{
A data frame with 1191565 rows and 7 columns
\describe{
\item{category}{Category}
\item{unit_nm}{Unit Name}
\item{GEOID}{GEOID}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.usgs.gov/core-science-systems/science-analytics-and-synthesis/gap/science/pad-us-data-overview?qt-science_center_objects=4#qt-science_center_objects}
}
\usage{
protected_areas
}
\description{
The dataset of land and waters in South Carolina which are dedicated to the preservation of biological diversity and available for natural,
recreational, and cultural purposes. These areas are protected from development and provide natural green spaces to the community. The
processed dateset includes each latitude/longitude point along the border of each area polygon, the protection category, area name, and census
information. The dataset is categorized as a Natural and Built Environment Determinant of Health.
}
\keyword{datasets}
