% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public_schools}
\alias{public_schools}
\title{South Carolina Public Schools}
\format{
A data frame with 1183 rows and 9 columns
\describe{
\item{ncessch}{NCES school ID}
\item{leaid}{Unique Agency Code}
\item{leanm05}{Operating Agency}
\item{schnam05}{School Name}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://nces.ed.gov/ccd/CCDLocaleCode.asp}
}
\usage{
public_schools
}
\description{
The dataset of South Carolina public elementary and secondary schools with locale details
}
\keyword{datasets}
