% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recreation}
\alias{recreation}
\title{South Carolina Recreation Areas}
\format{
A data frame with 1687 rows and 8 columns
\describe{
\item{objectid}{Object ID}
\item{name}{Name}
\item{feattype}{Feature Type}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
}
}
\source{
\url{https://www.arcgis.com/home/item.html?id=4f14ea9215d1498eb022317458437d19}
}
\usage{
recreation
}
\description{
The dataset of recreational areas in South Carolina
}
\keyword{datasets}
