% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_walkability_index}
\alias{sc_walkability_index}
\title{South Carolina Walkability Index}
\format{
A data frame with 3059 rows and 3 columns
\describe{
\item{geoid20}{Census Block Group 12-digit FIPS Code (2018)}
\item{nt_wlk_in}{National Walkability Index Score}
\item{geometry}{Geometry}
}
}
\source{
\url{https://edg.epa.gov/metadata/catalog/search/resource/details.page?uuid=\%7B251AFDD9-23A7-4068-9B27-A3048A7E6012\%7D}
}
\usage{
sc_walkability_index
}
\description{
The dataset of measures from the US EPA's National Walkability Index for SC block groups
}
\keyword{datasets}
