% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{social_vulnerability_index}
\alias{social_vulnerability_index}
\title{CDC/ATSDR Social Vulnerability Index}
\format{
A data frame with 1103 rows and 4 columns
\describe{
\item{geoid}{GEOID}
\item{svi_value}{Social Vulnerability Value}
\item{tract_latitude}{Latitude of the Census Tract Centroid}
\item{tract_longitude}{Longitude of the Census Tract Centroid}
}
}
\source{
\url{https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html}
}
\usage{
social_vulnerability_index
}
\description{
The dataset of social vulnerability for SC census tract groups according to a CDC/ATSDR determination. The dataset was originally created to
identify areas in the United States that would likely need extra assistance in the case of hazardous situations and natural disasters so as
to provide more adequate care in those situations. However, social vulnerability can also be used as a determinant of health and assist in
the understanding of health disparities in the US. The social vulnerability index considers a wide array of factors including: poverty level,
household composition and size, minority status, language spoken, housing type, and vehicle access. These factors were compiled into a single
score of social vulnerability. The processed dataset includes the census tract GEOID, the social vulnerability score, and census information.
The dataset is categorized as a Social and Community Context Determinant of Health.
}
\keyword{datasets}
