% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tri}
\alias{tri}
\title{SC Toxic Release Inventory Form R and A Submissions}
\format{
A data frame with 59834 rows and 17 columns
\describe{
\item{chem_name}{Chemical Released}
\item{facility_name}{Facility Name}
\item{tri_chem_id}{TRI Chemical ID}
\item{tri_facility_id}{TRI Facility ID}
\item{naics_codes}{NAICS Code}
\item{total_off_site_release}{Total Pounds of Off-Site Toxic Chemical Release}
\item{total_on_off_site_release}{Total Pounds of On- and Off-Site Toxic Chemical Release}
\item{total_on_site_release}{Total Pounds of On-Site Toxic Chemical Release}
\item{total_production_related_waste}{Total Pounds of Toxic Chemicals Released due to Production}
\item{land_total_release}{Total Pounds of Toxic Chemicals Released to Land}
\item{air_total_release}{Total Pounds of Toxic Chemicals Released to Air}
\item{water_total_release}{Total Pounds of Toxic Chemicals Released to Water}
\item{geoid}{GEOID}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Trac Centroid Longitude}
\item{lat}{Latitude}
\item{long}{Longitude}
}
}
\source{
\url{https://enviro.epa.gov/facts/tri/form_ra_download.html}
}
\usage{
tri
}
\description{
The dataset of sites in South Carolina that have submitted a Toxic Release Inventory Form R or A
}
\keyword{datasets}
