% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicle_availability}
\alias{vehicle_availability}
\title{Percent of Housing Units with No Vehicle Access}
\format{
A data frame with 1103 rows and 9 columns
\describe{
\item{state_fips}{State FIPS Code}
\item{state}{State}
\item{geoid}{GEOID}
\item{census_tract_2}{County and Census Tract}
\item{year}{Data Collection Year}
\item{value}{Percentage Value}
\item{data_comment}{NA}
\item{tract_latitude}{Tract Centroid Latitude}
\item{tract_longitude}{Tract Centroid Longitude}
}
}
\source{
\url{https://ephtracking.cdc.gov/DataExplorer/}
}
\usage{
vehicle_availability
}
\description{
The dataset of South Carolina census tracts with percentage estimates of housing units with no vehicle access
}
\keyword{datasets}
